+++ lib/mb_admin/mb_admin.dart
@@ -1,6 +1,9 @@
+import 'package:mburger/mb_admin/mb_admin_visibility_settings.dart';
+
 import '../mb_manager.dart';
-import 'mb_multipart_form.dart';
-import 'mb_uploadable_element.dart';
+import 'mb_admin_push_settings.dart';
+import 'uploadable_elements/mb_multipart_form.dart';
+import 'uploadable_elements/mb_uploadable_element.dart';
 import 'package:http/http.dart' as http;
 
 /// The manager of the MBAdmin section, you will use this class to create, edit or delete data in MBurger.
@@ -18,33 +21,33 @@ class MBAdmin {
   /// - Parameters:
   ///   - [blockId]: The id of the block to add a section.
   ///   - [elements]: The elements of the section that will be created.
+  ///   - [visibilitySettings]: This property will tell MBurger the visibility settings for the section.
+  ///   - [pushSettings]: This property will tell MBurger if it should send a push notification when the section is published.
   /// - Returns a [Future] that completes when the section is created correctly.
   Future<void> addSectionToBlock(
     int blockId,
-    List<MBUploadableElement> elements,
-  ) async {
+    List<MBUploadableElement> elements, {
+    MBAdminPushSettings pushSettings,
+    MBAdminVisibilitySettings visibilitySettings,
+  }) async {
     String apiName = 'api/blocks/' + blockId.toString() + '/sections';
 
     var uri = Uri.https(MBManager.shared.endpoint, apiName);
     var request = http.MultipartRequest('POST', uri);
+
     for (MBUploadableElement element in elements) {
       List<MBMultipartForm> forms = element.toForm();
-      if (forms != null) {
-        for (MBMultipartForm form in forms) {
-          if (!form.isFile) {
-            request.fields[form.name] = form.value;
-          } else {
-            request.files.add(
-              await http.MultipartFile.fromPath(
-                form.name,
-                form.path,
-                contentType: form.mimeType,
-              ),
-            );
-          }
-        }
-      }
+      await _addMultipartFormsToRequest(request, forms);
     }
+    if (pushSettings != null) {
+      List<MBMultipartForm> pushForms = pushSettings.toForm();
+      await _addMultipartFormsToRequest(request, pushForms);
+    }
+    if (visibilitySettings != null) {
+      List<MBMultipartForm> visibilityForms = visibilitySettings.toForm();
+      await _addMultipartFormsToRequest(request, visibilityForms);
+    }
+
     request.headers.addAll(await MBManager.shared.headers());
     http.StreamedResponse response = await request.send();
     final responseString = await response.stream.bytesToString();
@@ -56,38 +59,60 @@ class MBAdmin {
   /// - Parameters:
   ///   - [sectionId]: The id of the section that needs to be edited.
   ///   - [elements]: The elements of the section that will be created.
+  ///   - [visibilitySettings]: This property will tell MBurger the visibility settings for the section.
+  ///   - [pushSettings]: This property will tell MBurger if it should send a push notification when the section is published.
   /// - Returns a [Future] that completes when the section is edited correctly.
   Future<void> editSection(
     int sectionId,
-    List<MBUploadableElement> elements,
-  ) async {
+    List<MBUploadableElement> elements, {
+    MBAdminPushSettings pushSettings,
+    MBAdminVisibilitySettings visibilitySettings,
+  }) async {
     String apiName = 'api/sections/' + sectionId.toString() + '/update';
 
     var uri = Uri.https(MBManager.shared.endpoint, apiName);
     var request = http.MultipartRequest('POST', uri);
+
     for (MBUploadableElement element in elements) {
       List<MBMultipartForm> forms = element.toForm();
-      if (forms != null) {
-        for (MBMultipartForm form in forms) {
-          if (!form.isFile) {
-            request.fields[form.name] = form.value;
-          } else {
-            request.files.add(
-              await http.MultipartFile.fromPath(
-                form.name,
-                form.path,
-                contentType: form.mimeType,
-              ),
-            );
-          }
-        }
-      }
+      await _addMultipartFormsToRequest(request, forms);
     }
+    if (pushSettings != null) {
+      List<MBMultipartForm> pushForms = pushSettings.toForm();
+      await _addMultipartFormsToRequest(request, pushForms);
+    }
+    if (visibilitySettings != null) {
+      List<MBMultipartForm> visibilityForms = visibilitySettings.toForm();
+      await _addMultipartFormsToRequest(request, visibilityForms);
+    }
+
     request.headers.addAll(await MBManager.shared.headers());
     http.StreamedResponse response = await request.send();
     final responseString = await response.stream.bytesToString();
 
     MBManager.checkResponse(responseString, checkBody: false);
+  }
+
+  /// Add multipart forms to the request
+  Future<void> _addMultipartFormsToRequest(
+    http.MultipartRequest request,
+    List<MBMultipartForm> forms,
+  ) async {
+    if (forms != null) {
+      for (MBMultipartForm form in forms) {
+        if (!form.isFile) {
+          request.fields[form.name] = form.value;
+        } else {
+          request.files.add(
+            await http.MultipartFile.fromPath(
+              form.name,
+              form.path,
+              contentType: form.mimeType,
+            ),
+          );
+        }
+      }
+    }
   }
 
   /// Deletes a section.