+++ lib/mb_admin/uploadable_elements/mb_uploadable_elements_factory.dart
@@ -1,17 +1,23 @@
 import 'package:http_parser/http_parser.dart';
-import 'package:mburger/mb_admin/mb_uploadable_dropdown_element.dart';
-import 'package:mburger/mb_admin/mb_uploadable_multiple_element.dart';
-import 'package:mburger/mb_admin/mb_uploadable_relation_element.dart';
 
-import 'package:mburger/mb_admin/mb_uploadable_checkbox_elelment.dart';
-import 'package:mburger/mb_admin/mb_uploadable_images_elements.dart';
-import 'package:mburger/mb_admin/mb_uploadable_text_element.dart';
+import 'mb_uploadable_dropdown_element.dart';
+import 'mb_uploadable_multiple_element.dart';
+import 'mb_uploadable_relation_element.dart';
 
+import 'mb_uploadable_checkbox_elelment.dart';
+import 'mb_uploadable_images_elements.dart';
+import 'mb_uploadable_text_element.dart';
+
+/// Used to create MBUploadableElement without specifiyng the locale for every item.
+/// The locale is initialized with the factory and passed to all the objects. You can also change the locale and use the same instance of a MBUploadableElementsFactory to create objects with a different locale.
 class MBUploadableElementsFactory {
+  /// The locale identifier passed to every objects created.
   final String localeIdentifier;
 
+  /// Initializes a factory with the locale identifier.
   MBUploadableElementsFactory(this.localeIdentifier);
 
+  /// Creates a text element with a [name] and some [text].
   MBUploadableTextElement createTextElement(
     String name,
     String text,
@@ -23,6 +29,7 @@ class MBUploadableElementsFactory {
     );
   }
 
+  /// Creates an image element with a [name], the [imagePath], and its [mediaType].
   MBUploadableImagesElement createImageElement(
     String name,
     String imagePath,
@@ -36,6 +43,7 @@ class MBUploadableElementsFactory {
     );
   }
 
+  /// Creates an images element with a [name], a list of paths ([imagePaths]), and their [mediaType].
   MBUploadableImagesElement createImagesElement(
     String name,
     List<String> imagePaths,
@@ -49,6 +57,7 @@ class MBUploadableElementsFactory {
     );
   }
 
+  /// Creates a checkbox element with a [name], a its [value].
   MBUploadableCheckboxElement createCheckboxElement(String name, bool value) {
     return MBUploadableCheckboxElement(
       localeIdentifier,
@@ -57,6 +66,7 @@ class MBUploadableElementsFactory {
     );
   }
 
+  /// Creates a relation element with the [name] of the element, and the [sectionIds].
   MBUploadableRelationElement createRelationElement(
     String name,
     List<int> sectionIds,
@@ -68,6 +78,7 @@ class MBUploadableElementsFactory {
     );
   }
 
+  /// Creates a dropdown element with the [name] of the element, and the [value] selected.
   MBUploadableDropdownElement createDropdownElement(
     String name,
     String value,
@@ -79,6 +90,7 @@ class MBUploadableElementsFactory {
     );
   }
 
+  /// Creates a multiple element with the [name] of the element, and the [values] selected.
   MBUploadableMultipleElement createMultipleElement(
     String name,
     List<String> values,
@@ -87,6 +99,15 @@ class MBUploadableElementsFactory {
       localeIdentifier,
       name,
       values,
+    );
+  }
+
+  /// Creates a slug element with the value of the [slug].
+  MBUploadableTextElement createSlugElement(String slug) {
+    return MBUploadableTextElement(
+      localeIdentifier,
+      'mburger_slug',
+      slug,
     );
   }
 }