+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## [0.0.1+23]
+## [0.0.2]
 
-* Added get section with slug.
+* Added the ability set the SLUG of a section.
+* Added the ability to configure the visibility of a section when creating/updating it.
+* Added the ability to configure the push notifications of a section when creating/updating it.